[kernel][lib][watchdog] Watchdog driver in kernel space
Add a hw watchdog driver and necessary headers and helper functions
to keep check on the functioning of cpus. This watchdog driver pets the
hw counter periodically. If there is a sw stuck up, then the bark handler
is called and eventually, the watchdog will bite and restart the cpus.

Signed-off-by: Prakruthi Deepak Heragu <pheragu@codeaurora.org>